+++ pages/Homepage/homepage.js
@@ -1,6 +1,7 @@
 import Head from 'next/head';
 import MetaDefaults from '../../components/MetaDefaults';
 import ReactMarkdown from 'react-markdown';
+import MarkdownLink from '../../components/MarkdownLink';
 import Loader from '../../components/Loader';
 
 const Homepage = ({ homepage }) => {
@@ -23,7 +24,10 @@ const Homepage = ({ homepage }) => {
         <section className='section content'>
           <div className='container'>
             <h2 className='title is-2 has-text-centered'>{homepage.header}</h2>
-            <ReactMarkdown children={homepage.content} />
+            <ReactMarkdown
+              children={homepage.content}
+              components={{ a: MarkdownLink }}
+            />
           </div>
         </section>
       </div>