+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+PR Add MarkdownLink component - 10/14/2021
+
+- Add MarkdownLink component
+
 PR Add form data functionality - 10/14/2021
 
 - Add api route and functionality to save form